# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScreenplayAnalyzeInputMimeType = typeof ScreenplayAnalyzeInputMimeType[keyof typeof ScreenplayAnalyzeInputMimeType];


export const ScreenplayAnalyzeInputMimeType = {
  'text/plain': 'text/plain',
  'application/pdf': 'application/pdf',
} as const;

export interface ScreenplayAnalyzeInput {
  /** @minLength 1 */
  fileName: string;
  mimeType: ScreenplayAnalyzeInputMimeType;
  /**
     * Plain text or base64-encoded PDF content
     * @minLength 1
     */
  content: string;
}

export type AnalysisJobStatus = typeof AnalysisJobStatus[keyof typeof AnalysisJobStatus];


export const AnalysisJobStatus = {
  queued: 'queued',
  analyzing: 'analyzing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface AnalysisJob {
  screenplayId: string;
  status: AnalysisJobStatus;
}

export type ScreenplaySummaryStatus = typeof ScreenplaySummaryStatus[keyof typeof ScreenplaySummaryStatus];


export const ScreenplaySummaryStatus = {
  queued: 'queued',
  analyzing: 'analyzing',
  ready: 'ready',
  approved: 'approved',
  rejected: 'rejected',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type ScreenplaySummaryRecommendation = typeof ScreenplaySummaryRecommendation[keyof typeof ScreenplaySummaryRecommendation] | null;


export const ScreenplaySummaryRecommendation = {
  Pass: 'Pass',
  Consider: 'Consider',
  Recommend: 'Recommend',
} as const;

export interface ScreenplaySummary {
  id: string;
  fileName: string;
  status: ScreenplaySummaryStatus;
  /** @nullable */
  recommendation?: ScreenplaySummaryRecommendation;
  createdAt: string;
}

export type TraceEventStatus = typeof TraceEventStatus[keyof typeof TraceEventStatus];


export const TraceEventStatus = {
  complete: 'complete',
  active: 'active',
  pending: 'pending',
  error: 'error',
} as const;

export interface TraceEvent {
  id: string;
  label: string;
  detail: string;
  status: TraceEventStatus;
  createdAt: string;
}

export type CoverageReportComparableTitlesItem = {
  title: string;
  reason: string;
};

export type CoverageReportRecommendation = typeof CoverageReportRecommendation[keyof typeof CoverageReportRecommendation];


export const CoverageReportRecommendation = {
  Pass: 'Pass',
  Consider: 'Consider',
  Recommend: 'Recommend',
} as const;

export interface CoverageReport {
  logline: string;
  synopsis: string;
  strengths: string[];
  weaknesses: string[];
  comparableTitles: CoverageReportComparableTitlesItem[];
  recommendation: CoverageReportRecommendation;
}

export type ScreenplayDetailStatus = typeof ScreenplayDetailStatus[keyof typeof ScreenplayDetailStatus];


export const ScreenplayDetailStatus = {
  queued: 'queued',
  analyzing: 'analyzing',
  ready: 'ready',
  approved: 'approved',
  rejected: 'rejected',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type ScreenplayDetailDecision = typeof ScreenplayDetailDecision[keyof typeof ScreenplayDetailDecision] | null;


export const ScreenplayDetailDecision = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ScreenplayDetail {
  id: string;
  fileName: string;
  mimeType: string;
  status: ScreenplayDetailStatus;
  createdAt: string;
  /** @nullable */
  pageCount?: number | null;
  report: CoverageReport | null;
  trace: TraceEvent[];
  /** @nullable */
  keyArtUrl: string | null;
  /** @nullable */
  decision: ScreenplayDetailDecision;
  /** @nullable */
  decisionAt?: string | null;
}

export type DecisionInputDecision = typeof DecisionInputDecision[keyof typeof DecisionInputDecision];


export const DecisionInputDecision = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface DecisionInput {
  decision: DecisionInputDecision;
}