# lib/api-zod/src/generated/types/analysisJob.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisJobStatus } from './analysisJobStatus';

export interface AnalysisJob {
  screenplayId: string;
  status: AnalysisJobStatus;
}