# lib/api-zod/src/generated/types/coverageReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoverageReportComparableTitlesItem } from './coverageReportComparableTitlesItem';
import type { CoverageReportRecommendation } from './coverageReportRecommendation';

export interface CoverageReport {
  logline: string;
  synopsis: string;
  strengths: string[];
  weaknesses: string[];
  comparableTitles: CoverageReportComparableTitlesItem[];
  recommendation: CoverageReportRecommendation;
}