# lib/api-zod/src/generated/types/decisionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionInputDecision } from './decisionInputDecision';

export interface DecisionInput {
  decision: DecisionInputDecision;
}