# lib/api-zod/src/generated/types/screenplayAnalyzeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScreenplayAnalyzeInputMimeType } from './screenplayAnalyzeInputMimeType';

export interface ScreenplayAnalyzeInput {
  /** @minLength 1 */
  fileName: string;
  mimeType: ScreenplayAnalyzeInputMimeType;
  /**
     * Plain text or base64-encoded PDF content
     * @minLength 1
     */
  content: string;
}