# lib/api-zod/src/generated/types/screenplayDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoverageReport } from './coverageReport';
import type { ScreenplayDetailDecision } from './screenplayDetailDecision';
import type { ScreenplayDetailStatus } from './screenplayDetailStatus';
import type { TraceEvent } from './traceEvent';

export interface ScreenplayDetail {
  id: string;
  fileName: string;
  mimeType: string;
  status: ScreenplayDetailStatus;
  createdAt: string;
  /** @nullable */
  pageCount?: number | null;
  report: CoverageReport | null;
  trace: TraceEvent[];
  /** @nullable */
  keyArtUrl: string | null;
  /** @nullable */
  decision: ScreenplayDetailDecision;
  /** @nullable */
  decisionAt?: string | null;
}