# lib/api-zod/src/generated/types/screenplaySummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScreenplaySummaryRecommendation } from './screenplaySummaryRecommendation';
import type { ScreenplaySummaryStatus } from './screenplaySummaryStatus';

export interface ScreenplaySummary {
  id: string;
  fileName: string;
  status: ScreenplaySummaryStatus;
  /** @nullable */
  recommendation?: ScreenplaySummaryRecommendation;
  createdAt: string;
}