# lib/api-zod/src/generated/types/traceEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TraceEventStatus } from './traceEventStatus';

export interface TraceEvent {
  id: string;
  label: string;
  detail: string;
  status: TraceEventStatus;
  createdAt: string;
}